+++ day2/src/GiftShop/GiftShop.h
@@ -6,8 +6,8 @@
 class GiftShop {
     public:
         GiftShop();
-        int solve(std::string& ranges);
-        int solve2(std::string& ranges);
+        long solve(std::string& ranges);
+        long solve2(std::string& ranges);
     private:
         
         std::vector<std::string> idStringHandler(std::string& ogString);
